reactror: extracts login page template [skip ci]
